requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
mysql-connector==2.2.9
mysql-connector-python==8.0.18
